102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 int main(void)
 {
 	int n = 50;
-	unsigned long long fib1 = 1, fib2 = 2, next;
+	unsigned long fib1 = 1, fib2 = 2, next;
 	int count = 0;
 
-	printf("%llu, %llu, ", fib1, fib2);
+	printf("%lu, %lu, ", fib1, fib2);
 for (count = 2; count < n; count++)
 {
 next = fib1 + fib2;
-printf("%llu", next);
+printf("%lu", next);
 
 if (count < n - 1)
 {
